Split setting message from config txn family
